\name{NipponMap-package}
\alias{NipponMap-package}
\alias{NipponMap}
\docType{package}
\title{Japanese Map Data and Functions}
\description{
Digital map data of Japan for choropleth mapping, including a circle cartogram.
}
\details{
\tabular{ll}{
Package: \tab NipponMap\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2017-11-29\cr
License: \tab GPL (>=2)\cr
}
To avoid troubles with Japanese strings, non-ASCII Japanese characters
in R object is sanitized or converted into ASCII characters by utility
functions in this package. Some common data for Japanese are planned to
preset, for preventing from burdensome input. (Those will be reinforced
in further version).
}

\author{Susumu Tanimura \email{aruminat@gmail.com}}

% \references{
% ~~ Literature or other references for background information ~~
% }
% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
% ~~ the R documentation directory ~~
%\keyword{packagelibrary}
%\keyword{character}
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% ~~ simple examples of the most important functions ~~
% }
