\name{NipponMap-package}
\alias{NipponMap-package}
\alias{NipponMap}
\docType{package}
\title{
\packageTitle{NipponMap}
}
\description{
\packageDescription{NipponMap}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NipponMap}
\packageIndices{NipponMap}
%% \tabular{ll}{
%% Package: \tab NipponMap\cr
%% Type: \tab Package\cr
%% Version: \tab 0.1\cr
%% Date: \tab 2017-11-29\cr
%% License: \tab GPL (>=2)\cr
%% }
The package provides digital map data and functions of Japan for very
simple choropleth mapping in prefecture level. The cartographic function
in \pkg{Nippon} package was moved to \pkg{NipponMap} package. If a user
needs more complex digital map data or in municipal level, such data is
available in other package (e.g., \pkg{jpndistrict} and
\pkg{divagis}). }

\author{
\packageAuthor{NipponMap}

Maintainer: \packageMaintainer{NipponMap}
}

%% \references{
%% ~~ Literature or other references for background information ~~
%% }
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~

%% ~~ the R documentation directory ~~
%% \keyword{ package }
%% \seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%% }
%% \examples{
%% ~~ simple examples of the most important functions ~~
%% }
